add remote ir function
