+++ settings.ini
@@ -11,6 +11,10 @@ ir_led = 19
 
 [FILE]
 out_file_name = data/result.csv
+codes_aircon_file = pigpios/codes_aircon
+remote_aircon_ir_file = remote_ir.py
+main_dir = /home/pi/python/raspi
+
 
 [SSH]
 host_0 = raspi0-1.local
